pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Olderestin <sdordiuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.2"
python-telegram-bot = "20.7"
pydantic-settings = "2.1.0"
loguru = "0.7.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
